--- Webcrawler/SteamWebCrawler/src/main/java/de/collectioncompanion/SteamWebCrawler/SteamWebCrawlerApplication.java
@@ -1,13 +1,0 @@
-package de.collectioncompanion.SteamWebCrawler;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class SteamWebCrawlerApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(SteamWebCrawlerApplication.class, args);
-    }
-
-}
