--- Webcrawler/SteamWebCrawler/src/main/java/de/collectioncompanion/SteamWebCrawler/ports/data_files/WebCrawler.java
@@ -1,15 +1,0 @@
-package de.collectioncompanion.SteamWebCrawler.ports.data_files;
-
-import ports.Collection;
-
-public interface WebCrawler {
-
-    /**
-     * Searches for a collection with passed search term -> API call, HTML checking, ...
-     *
-     * @param searchTerm The search term to look for a collection
-     * @return The requested collection
-     */
-    Collection findInformationToCollection(String searchTerm);
-
-}
