--- Webcrawler/SteamWebCrawler/src/main/java/de/collectioncompanion/SteamWebCrawler/ports/outbound/GameOut.java
@@ -1,27 +1,0 @@
-package de.collectioncompanion.SteamWebCrawler.ports.outbound;
-
-import ports.Collection;
-
-import java.util.TreeMap;
-
-/**
- * Can only look for games -> category = games
- */
-public interface GameOut {
-
-    /**
-     * Request all available games on steam as "app-ID"-"game name" pair
-     *
-     * @return A TreeMap of all "app-ID"-"game name" pairs
-     */
-    TreeMap<Integer, String> getAllGames();
-
-    /**
-     * Searches for a collection with passed search term -> Steam-API call
-     *
-     * @param searchTerm The search term to look for a collection
-     * @return The requested collection
-     */
-    Collection findInformationToCollection(String searchTerm);
-
-}
