--- Webcrawler/SteamWebCrawler/src/main/java/de/collectioncompanion/SteamWebCrawler/ports/outbound/RAWGioOut.java
@@ -1,5 +1,0 @@
-package de.collectioncompanion.SteamWebCrawler.ports.outbound;
-
-public interface RAWGioOut extends GameOut {
-    
-}
