--- Webcrawler/SteamWebCrawler/src/main/java/de/collectioncompanion/SteamWebCrawler/ports/outbound/SteamAPI.java
@@ -1,5 +1,0 @@
-package de.collectioncompanion.SteamWebCrawler.ports.outbound;
-
-public interface SteamAPI extends GameOut {
-
-}
