--- Webcrawler/SteamWebCrawler/src/test/java/de/collectioncompanion/SteamWebCrawler/SteamWebCrawlerApplicationTests.java
@@ -1,13 +1,0 @@
-package de.collectioncompanion.SteamWebCrawler;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class SteamWebCrawlerApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
