+++ Webcrawler/src/main/java/de/collectioncompanion/WebCrawler/WebCrawlerApplication.java
@@ -1,0 +1,13 @@
+package de.collectioncompanion.WebCrawler;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class WebCrawlerApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(WebCrawlerApplication.class, args);
+    }
+
+}
