+++ Webcrawler/src/main/java/de/collectioncompanion/WebCrawler/adapter/outbound/RAWGioAPIOut.java
@@ -1,0 +1,19 @@
+package de.collectioncompanion.WebCrawler.adapter.outbound;
+
+import de.collectioncompanion.WebCrawler.ports.data_files.WebCrawler;
+import de.collectioncompanion.WebCrawler.ports.outbound.RAWGioOut;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import ports.Collection;
+
+@Service
+public class RAWGioAPIOut implements WebCrawler {
+
+    @Autowired
+    private RAWGioOut rawGioOut;
+
+    public Collection findInformationToCollection(String searchTerm) {
+        return rawGioOut.findInformationToCollection(searchTerm);
+    }
+
+}
