+++ Webcrawler/src/main/java/de/collectioncompanion/WebCrawler/ports/outbound/RAWGioOut.java
@@ -1,0 +1,5 @@
+package de.collectioncompanion.WebCrawler.ports.outbound;
+
+public interface RAWGioOut extends GameOut {
+    
+}
