+++ Webcrawler/src/main/java/de/collectioncompanion/WebCrawler/ports/outbound/SteamAPI.java
@@ -1,0 +1,5 @@
+package de.collectioncompanion.WebCrawler.ports.outbound;
+
+public interface SteamAPI extends GameOut {
+
+}
