+++ Webcrawler/src/test/java/de/collectioncompanion/WebCrawler/WebCrawlerApplicationTests.java
@@ -1,0 +1,13 @@
+package de.collectioncompanion.WebCrawler;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class WebCrawlerApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
